Refactor: 메인페이지 보완2

+++ src/main/java/com/hotsix/iAmNotAlone/domain/main/model/dto/PostProjection.java
@@ -20,7 +20,7 @@ public interface PostProjection {
 
     String getNickName();
 
-    int getGender();
+    Long getGender();
 
     String getUserFile();
